+++ R/votedates.R
@@ -23,28 +23,40 @@
 
 available_votedates <- function(geolevel="national"){
   
+  # datum aus coverage attribut auslesen -> urls$result$resources$coverage
+  
   if(!geolevel %in% c("national","canton")) stop("geolevel must be set to either 'national' or 'canton'")
 
-  #national votedates
+  #add option for cantonal votedates
   if(geolevel=="national"){
     
   urls <- jsonlite::fromJSON("https://opendata.swiss/api/3/action/package_show?id=echtzeitdaten-am-abstimmungstag-zu-eidgenoessischen-abstimmungsvorlagen")
   
-  # get votedates from dcat coverage attribute
+  # dates <- substr(urls$result$resources$name$de,21,30)
   
-  dates <- urls$result$resources$coverage
  
   }
   
-  #cantonal votedates
+  
   if(geolevel=="canton"){
     
   urls <- jsonlite::fromJSON("https://opendata.swiss/api/3/action/package_show?id=echtzeitdaten-am-abstimmungstag-zu-kantonalen-abstimmungsvorlagen")
     
-  dates <- urls$result$resources$coverage
+    
+  # dates <- substr(urls$result$resources$name$de,21,30)
     
   }
   
-  as.Date(dates)
+  dates <- urls$result$resources$coverage
+  
+
+  
+# Convert to Date Format - fallback if coverage attribute contains no information: get date from resource title 
+tryCatch(as.Date(dates), error=function(e) as.Date(substr(urls$result$resources$name$de,21,30),format="%d.%m.%Y"))
+  
+  
+  # as.Date(dates,format="%d.%m.%Y")
+  
+  # as.Date(dates)
   
 }