+++ tests/testthat/test_canton_json_to_dfr.R
@@ -1,0 +1,11 @@
+context("canton_json_to_dfr")
+
+test_that("canton_json_to_dfr works", {
+  
+  expect_error(canton_json_to_dfr(votedates="1999-01-01vv"))
+  
+  cantvotes <- canton_json_to_dfr()
+  
+  expect_is(cantvotes, "data.frame")
+  
+})